+++ src/main/java/com/starkbank/Transaction.java
@@ -157,6 +157,7 @@ public final class Transaction extends Resource {
      * before [string, default null] date filter for objects created only before specified date. ex: "2020-03-10"
      * tags [list of strings, default null]: tags to filter retrieved objects. ex: ["tony", "stark"]
      * externalIds [list of strings, default null]: list of external ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
+     * ids [list of strings, default null]: list of ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
      * <p>
      * Return:
      * @return generator of Transaction objects with updated attributes
@@ -206,6 +207,7 @@ public final class Transaction extends Resource {
      * after [string, default null] date filter for objects created only after specified date. ex: "2020-03-10"
      * before [string, default null] date filter for objects created only before specified date. ex: "2020-03-10"
      * externalIds [list of strings, default null]: list of external ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
+     * ids [list of strings, default null]: list of ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
      * @param user [Project object, default null]: Project object. Not necessary if starkbank.User.defaultUser was set before function call
      * <p>
      * Return: