+++ src/main/java/com/starkbank/Transfer.java
@@ -204,6 +204,7 @@ public final class Transfer extends Resource {
      * taxId [string, default null]: filter for transfers sent to the specified tax ID. ex: "012.345.678-90"
      * sort [string, default "-created"]: sort order considered in response. Valid options are "created", "-created", "updated" or "-updated".
      * tags [list of strings, default null]: tags to filter retrieved objects. ex: ["tony", "stark"]
+     * ids [list of strings, default null]: list of ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
      * <p>
      * Return:
      * @return generator of Transfer objects with updated attributes
@@ -256,6 +257,7 @@ public final class Transfer extends Resource {
      * status [string, default null]: filter for status of retrieved objects. ex: "processing" or "success"
      * sort [string, default "-created"]: sort order considered in response. Valid options are "created", "-created", "updated" or "-updated".
      * tags [list of strings, default null]: tags to filter retrieved objects. ex: ["tony", "stark"]
+     * ids [list of strings, default null]: list of ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
      * @param user [Project object, default null]: Project object. Not necessary if starkbank.User.defaultUser was set before function call
      * <p>
      * Return: