+++ CHANGELOG.md
@@ -13,6 +13,12 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- ids parameter to Transaction.query
+- ids parameter to Transfer.query
+
+### Fixed
+- Fixed params alterations on queries
 
 ## [0.6.0] - 2020-08-19
 ### Added